\name{hpvP53}
\alias{hpvP53}
\docType{data}
\title{ Time-course P53 pathway data }
\description{
Time-course gene expression data of the P53-signalling pathway during HPV-induced transformation.
}
\usage{data(hpvP53)}
\format{
An object of class \code{\link[Biobase:ExpressionSet]{ExpressionSet}}.
}
\source{
Wilting, S.M., Miok, V., Jaspers, A., Boon, D., Hanne Sorgard, H., Lando, M., Snoek, B.C., Van Wieringen, W.N., Meijer, C.J.L.M., Lyng, H., Snijders, P.J.F., Steenbergen, R.D.M. (2016), "Aberrant methylation-mediated silencing of microRNAs contributes to HPV-induced anchorage independence", , \emph{Oncotarget}, Vol. 7, No. 28, 43805-43819, DOI:10.18632/oncotarget.9698 .
}
\examples{
data(hpvP53)
}
\keyword{datasets}


